% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_scores.R
\name{calculate_game_score}
\alias{calculate_game_score}
\title{Calculate game score from a game result}
\usage{
calculate_game_score(winner_score, loser_score)
}
\arguments{
\item{winner_score}{The score of the winning team.}

\item{loser_score}{The score of the losing team.}
}
\description{
This function calculates game score using the USA Ultimate Rankings Algorithm.
It returns the game score for the winning team.
}
\examples{
calculate_game_score(13,12)
calculate_game_score(13,6)
}
