% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{calculate_game_score_adjusted_team}
\alias{calculate_game_score_adjusted_team}
\title{Calculate team rating change from a game result}
\usage{
calculate_game_score_adjusted_team(
  winner_team,
  loser_team,
  winner_score,
  loser_score,
  league_type
)
}
\arguments{
\item{winner_team}{The name of the winning team.}

\item{loser_team}{The name of the losing team.}

\item{winner_score}{The score of the winning team.}

\item{loser_score}{The score of the losing team.}

\item{league_type}{The league type of the game. Should be equal to "mens" for men's games and "womens" for women's games. Other values will return an error.}
}
\description{
This function calculates game score using the USA Ultimate Rankings Algorithm.
Then, it determines the ratings changes as a result of that game outcome.
Unlike its counterpart, calculate_game_score_adjusted, it takes team names — not ratings — as inputs.
Then, it pulls that team's most updated rating from the rankings site and calculates the results.
Returns a dataframe with information.
}
\examples{
calculate_game_score_adjusted_team("Virginia","Virginia Tech",13,6,league_type = "mens")
}
