% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_win_probability.R
\name{calculate_win_probability}
\alias{calculate_win_probability}
\title{Get pregame win probability}
\usage{
calculate_win_probability(TeamRating, OpponentRating, LeagueType)
}
\arguments{
\item{TeamRating}{The rating of the first team playing in the game. The order of TeamRating and OpponentRating is interchangeable.}

\item{OpponentRating}{The rating of the second team playing in the game. The order of TeamRating and OpponentRating is interchangeable.}

\item{LeagueType}{The league type of the game. Should be equal to "mens" for men's games and "womens" for women's games. Other values will return an error.}
}
\description{
This function returns the pregame win probability for a pair of teams, given each team's rating and the league type.
}
\examples{
calculate_win_probability(1500,1400,"mens")
calculate_win_probability(1500,1400,"womens")
}
