% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{GameScoreCalculator}
\alias{GameScoreCalculator}
\title{Calculate game score from a game result}
\usage{
GameScoreCalculator(winner_score, loser_score)
}
\arguments{
\item{winner_score}{The score of the winning team.}

\item{loser_score}{The score of the losing team.}
}
\description{
This function calculates game score using the USA Ultimate Rankings Algorithm.
It returns the game score for the winning team.
}
\examples{
GameScoreCalculator(13,12)
GameScoreCalculator(13,6)
}
