% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamesW}
\alias{gamesW}
\title{Game-level data for the college women's 2021-22 season.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://www.frisbee-rankings.com/}
}
\usage{
gamesW
}
\description{
A dataset containing information on each of the 4,220 documented women's ultimate frisbee games played at the collegiate level in the 2021-22 season.
}
\keyword{datasets}
