% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{GetFrisbeeRankings}
\alias{GetFrisbeeRankings}
\title{Get updated college men's frisbee rankings}
\usage{
GetFrisbeeRankings(DivisionIOnly = F, SimpleTable = F)
}
\arguments{
\item{DivisionIOnly}{Returns just D-I rankings if TRUE. Defaults to FALSE.}

\item{SimpleTable}{Returns just a table with team info & rating if TRUE. Defaults to FALSE.}
}
\description{
This function returns up-to-date ultimate frisbee rankings from frisbee-rankings.com
}
\examples{
cat_function()
}
\keyword{cats}
