% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{GetTeamResultsMen}
\alias{GetTeamResultsMen}
\title{Get a team's game results}
\usage{
GetTeamResultsMen(team)
}
\arguments{
\item{team}{The name of the team you want to scrape data for.}
}
\description{
This function scrapes a team's game results and computes some useful info for each game.
}
\examples{
GetTeamResultsMen("Virginia")
}
