% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glossary_AUDL_teams}
\alias{glossary_AUDL_teams}
\title{Glossary containing AUDL team names and logos}
\format{
A data frame with 25 rows and 4 variables:
\describe{
  \item{teamCity}{The team's home city.}
  \item{teamName}{The team's nickname.}
  \item{teamAbbr}{The team's two- or three-letter abbrevation.}
  \item{logoURL}{URL of the team's logo.}
}
}
\source{
URLs come from a variety of sources: mostly the AUDL website, but when logos were unavailable there, they were accessed via Wikipedia or Shopify team site.s
}
\usage{
glossary_AUDL_teams
}
\description{
A win probability model fit in base R using logistic regression to the modelW dataset with one feature: Elo difference.
}
\keyword{datasets}
