% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{load_audl_games}
\alias{load_audl_games}
\title{Load AUDL games}
\usage{
load_audl_games(seasons)
}
\arguments{
\item{seasons}{The seasons for which data is requested. All seasons since 2012 are documented, not including 2020 when the AUDL season was cancelled due to the COVID-19 pandemic.}
}
\description{
Loads information on all AUDL games from the specified season into memory, scraped from the AUDL website.
}
\examples{
load_audl_games(2022)
load_audl_games(c(2019,2021,2022))

}
