% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audl_scraping.R
\name{load_audl_player_stats}
\alias{load_audl_player_stats}
\title{Load AUDL player statistics}
\usage{
load_audl_player_stats(season, stat_type = "game")
}
\arguments{
\item{season}{The season for which data is requested. All seasons since 2012 are documented, not including 2020 when the AUDL season was cancelled due to the COVID-19 pandemic.}

\item{stat_type}{The desired statistic type. Set to "game" by default, for per-game statistics. Can also be set to "total" for total statistics, "10 points" for per-10-point statistics, or "10 possessions" for per-10-possession statistics.}
}
\description{
Loads AUDL player statistics from the specified season into memory, scraped from the AUDL website.
}
\examples{
load_audl_player_stats(2022)
load_audl_player_stats(2021, stat_type = "10 points")
}
