% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{load_audl_team_stats}
\alias{load_audl_team_stats}
\title{Load AUDL team statistics}
\usage{
load_audl_team_stats(season, stat_type = "game", team_type = "team")
}
\arguments{
\item{season}{The season for which data is requested. All seasons since 2012 are documented, not including 2020 when the AUDL season was cancelled due to the COVID-19 pandemic.}

\item{stat_type}{The desired statistic type. Set to "game" by default, for per-game statistics. Can also be set to "total" for total statistics.}

\item{team_type}{Which team the returned statistics should describe. "team" is the default, and will return the statistics for each team. "opponent" will return statistics for opponent performance against each team.}
}
\description{
Loads AUDL team statistics from the specified season into memory, scraped from the AUDL website.
}
\examples{
load_audl_team_stats(2022)
load_audl_team_stats(2022, stat_type = "total", team_type = "opponent")
}
