% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_scraping_high_level.R
\name{load_rankings_men}
\alias{load_rankings_men}
\title{Get updated college men's frisbee rankings}
\usage{
load_rankings_men(DivisionIOnly = F, SimpleTable = F)
}
\arguments{
\item{DivisionIOnly}{Returns just D-I rankings if TRUE. Defaults to FALSE.}

\item{SimpleTable}{Returns just a table with team info & rating if TRUE. Defaults to FALSE.}
}
\description{
This function returns up-to-date ultimate frisbee rankings from frisbee-rankings.com
}
\examples{
load_rankings_men()
}
