% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{load_rankings_women}
\alias{load_rankings_women}
\title{Get updated college women's frisbee rankings}
\usage{
load_rankings_women(DivisionIOnly = F, SimpleTable = F)
}
\arguments{
\item{DivisionIOnly}{Returns just D-I rankings if TRUE. Defaults to FALSE.}

\item{SimpleTable}{Returns just a table with team info & rating if TRUE. Defaults to FALSE.}
}
\description{
This function returns up-to-date ultimate frisbee rankings from frisbee-rankings.com
}
\examples{
load_rankings_women()
}
