% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_scraping_team_level.R
\name{load_team_results_women}
\alias{load_team_results_women}
\title{Get a team's game results}
\usage{
load_team_results_women(team)
}
\arguments{
\item{team}{The name of the team you want to scrape data for.}
}
\description{
This function scrapes a team's game results and computes some useful info for each game.
}
\examples{
load_team_results_women("Virginia")
}
