% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logitmodelM}
\alias{logitmodelM}
\title{Win probability model for men's college Ultimate Frisbee}
\format{
An R generalized linear model object.
}
\source{
\url{https://www.frisbee-rankings.com/}
}
\usage{
logitmodelM
}
\description{
A win probability model fit in base R using logistic regression to the modelM dataset with one feature: Elo difference.
}
\keyword{datasets}
