% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logitmodelW}
\alias{logitmodelW}
\title{Win probability model for women's college Ultimate Frisbee}
\format{
An R generalized linear model object.
}
\source{
\url{https://www.frisbee-rankings.com/}
}
\usage{
logitmodelW
}
\description{
A win probability model fit in base R using logistic regression to the modelW dataset with one feature: Elo difference.
}
\keyword{datasets}
