% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultimate-Frisbee-Functions.R
\name{RatingAdjustedGameScoreCalculatorTeam}
\alias{RatingAdjustedGameScoreCalculatorTeam}
\title{Calculate team rating change from a game result}
\usage{
RatingAdjustedGameScoreCalculatorTeam(
  winner_team,
  loser_team,
  winner_score,
  loser_score,
  input_data = suppressWarnings(GetFrisbeeRankings())
)
}
\arguments{
\item{winner_team}{The name of the winning team.}

\item{loser_team}{The name of the losing team.}

\item{winner_score}{The score of the winning team.}

\item{loser_score}{The score of the losing team.}
}
\description{
This function calculates game score using the USA Ultimate Rankings Algorithm.
Then, it determines the ratings changes as a result of that game outcome.
Unlike its counterpart, RatingAdjustedGameScoreCalculator, it takes team names — not ratings — as inputs.
Then, it pulls that team's most updated rating from the rankings site and calculates the results.
Returns a dataframe with information.
}
\examples{
RatingAdjustedGameScoreCalculatorTeam("Virginia","Virginia Tech",13,6)
}
